Add Inital spawn position of player.
This pull request introduces the functionality of setting an initial spawn point for the player character. A `Marker3D` node was added in the castle scene, which is leveraged as the starting point for the player's position when the game begins. In the player.gd scope.

I dont know if the uid thing is a problem for other editior, i'll modify the pr if needed.