Implement Initial Door Animation and Logic for Game Loading
Description:
In this pull request, I implemented the initial door animation and logic for when the game is first loaded. This involves creating the animation for the door, setting up the logic to control the animation, and utilizing the existing SceneSwitcher script to ensure the MainMenu remains easy to understand and to facilitate the interaction between the two scripts. Additionally, I created a script for the World scene, which can be used to globalize other states in the future. This PR also includes the implementation of the initial spawn position when switching to the inside of the castle, currently using a global state for this purpose, and fixing a bug that caused the game to crash when changing scenes.